MB-6074: Fix observe unit test failures on Ubuntu
These unit test failures were caused because waiting for
the flusher queue to drain means that the item is on disk
but the StoredValue data has not yet been updated. This
update happens very shortly after the item hits disk, but
in our test cases this cause caused sporadic failures if
we called observe before the StoredValue was updated. The
fix for this is to check the stat for total persisted items
instead.

Change-Id: I82eb544d3253567e345ce1bfae2cccb1ed772b67
Reviewed-on: http://review.couchbase.org/19132
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Reviewed-by: Jin Lim <jin@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>