MB-6075: Fix xxx-with-meta sporadic failures
These sporadic failures were caused because of we were
waiting for the flusher. The problem is that the flusher
draining means that the items hit disk and does not imply
that the StoredValue for each persisted object was updated.
In the case of delete the callback decrements curr_items
and this caused some stat checks to fail incorrectly. This
cahnge makes sure we wait for everything to be updated
correctly before checking the stats.

Change-Id: Ie7244101443cfaee492083300754affb367ff72b
Reviewed-on: http://review.couchbase.org/19136
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>