MB-6062: Better error checking in couchkv-store getPersistedStats
This fix checks to see if we have recieved an error code from the
tellg ifstream function and logs an error and returns from the
function if so. We also add a NULL termination character to the end
of our ifstream buffer just in case one doesn't exist. This prevents
an error when parsing this value as json and adding the extra byte
prevents us from having to copy the entire buffer if a NULL
terminator doesn't already exist.

Change-Id: I08fdb6476d633d3a7a36e133bb11b601974cb60c
Reviewed-on: http://review.couchbase.org/19465
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>