MB-6323 Ignore EINVAL from pthread_mutex/cond_destroy()
On windows mingw environment, there are sporadic failures
in destroying a mutex in Mutex deconstructor. The culprit for
this issue on windows is still not identified. Ignoring EINVAL
returned from destroying a mutex doesn't cause any other issues
on windows at this time, but we will continue to investigate
this issue.

Change-Id: I6da20dc5ac0623d1931a443617a5663b6e053e55
Reviewed-on: http://review.couchbase.org/20322
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>