MB-6518: Make sure exptime is always persisted properly
This commit makes two changes. First it makes the fetchValidValue
function return a StoredValue, even in the case of an expired
value. This is needed because if we try to persist an expired
item an fetchValidValue returns NULL, then the flusher will
persist 0 as the value for flags and expiration time and also
persist a random cas value.

The second part of this change adds a queueExpired parameter to
the fetchValidValue function. This value is needed by the flusher
because if we recieve a op_set and the value is expired then we
do the delete in the flusher and also queue an op_del for the same
item. When the flusher gets to the op_del the item no longer exists
and as a result 0 is persisted as the exp time and flags time and
a random cas is persisted for that item.

Change-Id: Ife69dd45ed490e9277f2074ef7780831e017881a
Reviewed-on: http://review.couchbase.org/20631
Tested-by: buildbot <build@couchbase.com>
Tested-by: Junyi Xie <junyi.couchbase@gmail.com>
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>