MB-7023 Fix to race conditions in vbucket delete / recreate
Due to the cluster rebalance, a given vbucket can be
removed and recreated in memory and disk in a short time.

In this case, the dirty items in the vbucket queue, which
belong to the deleted vbucket should be simply skipped.
In addition, persisting incoming dirty items, which belong to
the newly created vbucket, should be blocked until the new
vbucket database is created on disk.

Change-Id: If733a9e758993066344a4ed3314fcb89bb354b24
Reviewed-on: http://review.couchbase.org/22001
Reviewed-by: Jin Lim <jin@couchbase.com>
Tested-by: Jin Lim <jin@couchbase.com>