MB-6992 Control the flusher execution by the transaction size
We recently removed the transaction size limit, which determines
how long the flusher can be executed by the RW thread without
being preemted by other higher priority tasks. However, this
allows the flusher to be run continuously without preemption
until its outgoing queue becomes empty. Consequently, other
high priority tasks such as vbucket deletion, vbucket snapshot,
should wait for a long time in the thread queue.

To resolve this issue, we bring back the transaction size limit
to the flusher, so that the flusher can yield the thread
when it reaches to the transaction size limit.

Change-Id: I15f2cb0b6855b1bc1553398db7c7e0349f2640c5
Reviewed-on: http://review.couchbase.org/22022
Reviewed-by: Jin Lim <jin@couchbase.com>
Tested-by: Jin Lim <jin@couchbase.com>