MB-6806 ep engine returns key not found for a deleted doc
This commit fixes the invalid error return (doc_not_found)
when the requested document is being marked deleted or when
the requested document is being compacted. In both cases,
ep engine now continues its mutation operation with the
default file revision number and let the callback handle the
error case with the key not found return instead.

Change-Id: I29c93eda060f16489830dc5d58f150ac1b0ec33a
Reviewed-on: http://review.couchbase.org/22023
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>