MB-7056 Don't check the deleted flag for bg getMeta() calls.
We invoke the underlying store's getMeta() to fetch the meta data
of a deleted item from disk. However, if the request is sitting
for a while in the RO dispatcher queue, there might be a case
where we receive a new insertion for the same key and persists
it into the vbucket database. In this case, the deleted flag
in the meta data is not true anymore.

Change-Id: If710e12276c9d6f326869de37d546663b544fda5
Reviewed-on: http://review.couchbase.org/22116
Reviewed-by: Jin Lim <jin@couchbase.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>