MB-7057 Schedule VB deletion first and snapshotVBs for resetVB()
resetVBucket() internally deletes a given vbucket and recreates
the empty vbucket in memory and disk. Therefore, the vbucket
deletion should be scheduled first and then the vbucket state
snapshot task be executed. Therefore, the flusher also should
skip the vbucket whose database creation request is still waiting
in the queue.

Change-Id: Ic936f5100f02d438367d9a1d50333818950a35d8
Reviewed-on: http://review.couchbase.org/22286
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Reviewed-by: Jin Lim <jin@couchbase.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>