MB-7471: Fix race condition in persistence checkpoint test
Thread 1 and 2 would issue a persist checkpoint command, both of
which are expected to time out. Thread 1's command would finish and
bump the timeout up and then go on to load items and call the
checkpoint persistence command again. This persistenc command would
finish and lower the time out value back to the default value. Thread
2's first checkpoint persistence command now finishes, but when we
check to see if the timeout was bumped up we fail the test.

We should just be checking that checkpoint persistence happens after
we join the threads in order to fix this problem.

Change-Id: Ie4b34e647684c2760a620e4c2aa40c3ea724caff
Reviewed-on: http://review.couchbase.org/23756
Reviewed-by: Jin Lim <jin@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>