MB-7519: Better flusher batching for writes
This change removes the outgoing queue in ep.cc and pushes
dirty items directly from the checkpoint queues into the
couchkvstore where they are persisted to disk. This
improves batching because we are now able to grab all
dirty items in a vbucket as opposed to grabbing items
an putting them into an intermediate queue where they
would wait to be persisted.

Change-Id: I39e9c25ad3380b3e045002967c9532b16758e187
Reviewed-on: http://review.couchbase.org/24190
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>