CBD-821 Extend an item's NRU value to two bits.
The current NRU scheme allocates only one bit per item in memory
cache and resets it to false whenever it is updated or referenced.
However, this scheme doesn't have the re-reference frequency of
each item and treats update and read operations with the same
weight in a cache management.

To support the easy and simple prediction of the re-reference
frequency, this change extends an item's NRU value to two bits,
so that it can have the re-reference hint for each item.

Basically, when a new item is inserted, its NRU value is set to
the intermediate value (i.e., two), and decremented as long as
the value is greater than zero whenever the item is referenced
later. The new item eviction algorithm (shown in another change)
uses an item's NRU value to collect the list of items with high
NRU values, which will be candidates for evictions.

Change-Id: Id863d68f1e0cfed6382a2ad53a2d7b089029df70
Reviewed-on: http://review.couchbase.org/24439
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>