MB-7590: Remove the mutation key log
In order to allow for hot items to be loaded back into
memory after the server is restarted, ep-engine keeps a
record of recently accessed/updated items. The bookkeeping
for this feature was implemented in two ways, the mutation
log and the access log. The mutation log records all
mutations into a file for the server to replay upon failure,
while the access log records the contents of the cache every
24 hours. The access log was chosen as the best way to
implement this feature since it requires less disk overhead. As
a result the mutation log has been turned off by default and
is not recommended for use. As a result this change removes the
mutation log in order to simplify the code base.

Change-Id: I8ffb708be6314ea3d715ce59cbf60d76c1bb264b
Reviewed-on: http://review.couchbase.org/26943
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>