MB-8497: Don't track alignment for stored value size function
This code was an attempt at making the sizeof memory accounting
more accurate. We don't use this type of accounting anymore and
this actually makes things much more difficult to keep track of
since we do not adjust for alignment everywhere. On top of this
not all memory allocations are done at 8 byte intervals so this
is technically not even correct code.

Change-Id: I2ddac215bab6803643123630dd146ccd4a49fc1e
Reviewed-on: http://review.couchbase.org/27035
Reviewed-by: Jin Lim <jin@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>