MB-8516: Change seqNum to revSeqno in QueuedItem class
We have two different sequence numbers that we will be working with
in ep-engine and it is very confusing to figure out which on we are
talking about when the variable name is specific. The xdcr sequence
number should be revSeqno when it is used in order to distiguish it
from the by sequence number.

Change-Id: I8e2a2764deb9ffc01d0dd43c9ee5d950196a88aa
Reviewed-on: http://review.couchbase.org/27123
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>