MB-8825 Use seq num and cas only for delete conflict resolution
Recently, we decided to use an expiration field for tombstone
purging, which can break the conflict resolution for the case
where the source and destination have the same seq number and
cas value. To resolve this issue, the change allow us to use
seq number and cas value only for delete conflict resolutions.

Change-Id: I51ac6eecc7e9f4499bd0ac7285f9021381fdaacd
Reviewed-on: http://review.couchbase.org/28398
Tested-by: Junyi Xie <junyi.couchbase@gmail.com>
Reviewed-by: Junyi Xie <junyi.couchbase@gmail.com>