MB-8869 Use atomic queue to push notifications for TAP producers
We saw the frontend ops/sec performance regression from the
previous fix to the OBSERVE replication latency issue. To address
this issue, this change uses the atomic queue to push
notification events for TAP producers, and schedules the
daemon task to grab those events and notify them to the
corresponding TAP producers.

Change-Id: I4bda29d50d446c338cff78014f20f2cef9f7bd64
Reviewed-on: http://review.couchbase.org/28464
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>