CBD-1097: GIO can Limit num of threads on a Queue
GIO Manager can now put a limit on the number of threads
which can pick up Tasks from specific queues.
This feature is currently used to ensure that at most 1 thread
ends up picking up tasks from the AuxIO queue so that we don't
end up in a case where all the threads are working
on the long running cron jobs of the AuxIO queue while there are
other higher priority tasks waiting to run.

Change-Id: I27ba74f0d9f3e0acf03c79604f36b26a53f60f83
Reviewed-on: http://review.couchbase.org/29136
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>