Only log message if we are acutally removing a tap cursor
There is a place in the code that calls this function for each
vbucket and this causes the logs to be flooded with messages
about removing tap cursors from vbuckets even when none are
removed. If we log the message after we check to see if the
cursor exists then the problem goes away.

Change-Id: Ie910422e7cca5e1da05a467f68aac78f6db12c53
Reviewed-on: http://review.couchbase.org/29572
Tested-by: buildbot <build@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>