MB-9179 Shard count made configurable
This change makes the number of shards configurable when a bucket
is instantiated and sets its default number to four. As part of
the global IO manager optimization, the workload policy (i.e.,
read-heavy, write-heavy, and mixed) parameter, which is used in
determining the number of shards, is no longer necessary.

As part of further enhancements, we plan to investigate
differentiating the number of shards depending on a bucket
priority.

Change-Id: I88f1dd0b681aaa562415c2748f8a0d9d9b1a26e3
Reviewed-on: http://review.couchbase.org/29234
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>