MB-8024: Fix to the BackfillRemaining stat overflow
Note that stats used in estimating the number of backfill items
are atomic individually, but not atomic when they use together.
Therefore, there is still a small time window that can cause the
backfill estimation counter to be overflow or underflow.

This change adds a check to prevent adding a negative number to
the backfill estimation counter.

Change-Id: I4752445e680c36b387aa80ffa1a985a7926bc9f2
Reviewed-on: http://review.couchbase.org/29836
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>
Tested-by: Chiyoung Seo <chiyoung@couchbase.com>