MB-8727: Cluster Configuration Carrier Publication
Addresses MB-8727:
- CMD_SET_CLUSTER_CONFIG: 0xb4

Addresses MB-8728:
- CMD_GET_CLUSTER_CONFIG: 0xb5

- Cluster config Blob will be saved as it is received
and will be returned as it is, with assumption that
blob will be of JSON type, and that SDK will need to
check the type of the document to differentiate
between older and new version of couchbase.

Addresses MB-8729:
- NOT_MY_VBUCKET to return cluster config blob
instead of the "not my vbucket" message.
- Memcached
change: http://review.couchbase.org/#/c/29844/
For memcached to retrieve Cluster Configuration from
ep_engine in case of ENGINE_NOT_MY_VBUCKET response.
- bucket_engine
change: http://review.couchbase.org/#/c/29845/

Change-Id: I20cf4aa145139d19524f252d3444dd0f52d0965d
Reviewed-on: http://review.couchbase.org/29684
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>
Tested-by: Chiyoung Seo <chiyoung@couchbase.com>