UPR consumer add_stream command
The upr add stream command is used in order to force a
consumer to initiate a connection with a producer. When
this command is received by a consumer it will initiate
an upr handshake over the same connection. Upon handshake
completion or failure a response for the add stream
command will be sent indicating the status  of the stream.

Change-Id: I40332293a9ae5df0c366297041b4d8d3fa871668
Reviewed-on: http://review.couchbase.org/29951
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>