Fix add stream engine api to fit protocol requirements
There were two issues here. The first is that when an add stream
is called we need to make the connection an upr connection so that
the the step api is invoked. The add stream api needs to send off
some other commands over the network and the step command allows
us to do this.

The second issue is that we do not want the add stream command to
return a response until we know a stream has been created. This
means that upon returning engine success we need to put the
connection in conn_ship_log state. The response for this api will
be sent at a later time during subsequent calls to the step api.

Change-Id: Ieadb8b792404f9802fe7b746b83f18ddb5d3741a
Reviewed-on: http://review.couchbase.org/30248
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>