MB-9521: Flusher must retry vbucket if no snapshot
When a bucket is created, ep-engine sets the flag
isBucketCreation into the vbMap for all its vbuckets.
These flags are only cleared by the the snapshot task
which should run first on all the vbuckets.
But if a flusher task runs first, then it will simply skip
that vbucket, Even if it has dirty items to flush!
Flusher should be able to retry a vbucket if it's
isBucketCreation flag is set so that it can run again
after the snapshot task is done.
Change-Id: Id41d5bf1dcc4e2459af5bb3719ff161f240c03d8
Reviewed-on: http://review.couchbase.org/30269
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>
Tested-by: Chiyoung Seo <chiyoung@couchbase.com>