MB-9532 Find a new compacted file upon receiving a temp failure
When the flusher receives a temp failure from the mccouch, it
should check if a new compacted file exists or not before
re-writing dirty items. This is required especially when the
deletion of the old file fails due to some external issues.
Otherwise, the flusher will keep writing dirty items into the
old file and receiving a temp failure over and over.

Change-Id: Ifd91eea1acfe82f5ea43ff93ca650b7928844b33
Reviewed-on: http://review.couchbase.org/30301
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Reviewed-by: Sundararaman Sridharan <sundar.sridharan@gmail.com>
Tested-by: Sundararaman Sridharan <sundar.sridharan@gmail.com>