MB-9203 Always reset the open checkpoint upon resetting a vbucket
When a replica vbucket is reset, its open checkpoint id is not
reset to zero when the vbucket has zero items. This causes a bug
in updating the persisted checkpoint id, which consequently
incurs the rebalance to be stuck.

This fix always resets the open checkpoint id of a replica
vbucket to zero upon resetting the vbucket.

Change-Id: If6b5e2ca9b37dd533818e8d3397ec2c695345295
Reviewed-on: http://review.couchbase.org/30302
Reviewed-by: Sundararaman Sridharan <sundar.sridharan@gmail.com>
Tested-by: Sundararaman Sridharan <sundar.sridharan@gmail.com>