CBD-1039 Delete expired non-resident items as part of compaction
This change allows us to delete an expired non-resident item
as part of compaction when the full eviction is selected.

For this, we create a temp item in the hash table and mark it
as deleted and then push the deletion into the checkpoint queue
for the replication and persistence.

Change-Id: I54ef0205179f75d9f8906e4dcacc8952b25bd4d5
Reviewed-on: http://review.couchbase.org/30396
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Chiyoung Seo <chiyoung@couchbase.com>