MB-9521: Flusher must retry vbucket if no snapshot
When a bucket is created, ep-engine sets the flag
isBucketCreation into the vbMap for all its vbuckets.
These flags are only cleared by the the snapshot task
which should run first on all the vbuckets.
But if a flusher task runs first, then it will simply skip
that vbucket, Even if it has dirty items to flush!
Flusher should be able to retry a vbucket if it's
isBucketCreation flag is set so that it can run again
after the snapshot task is done.

Change-Id: Ib92dd43c8fcc13c9076ed01d68de7735b20a079b
Reviewed-on: http://review.couchbase.org/30426
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>
Tested-by: Chiyoung Seo <chiyoung@couchbase.com>