Fix sporadic tap unit test failures
There were two issues here. The first is that we must wait for the
persistence cursor to move into the open checkpoint for the estimate
vb move test. This is because during checkpoint registration we now
place the cursor in the checkpoint that contains the persistence
cursor as opposed to just putting it into the open checkpoint. The
second issue was with the way we were calculating the cursor offset.

Change-Id: I48afb516479c1a19e52215fb4ab46de057fc46fa
Reviewed-on: http://review.couchbase.org/31582
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>