Reduce the number of sets for checkpoint test
The test takes like 22 minutes to run on windows with the
default number of inserts. By reducing the number to 10k
it takes 42secs. This patch reduce the number to 10k,
but we should back out this patch once we've figured out
why it runs so much slower on windows

Change-Id: I024bfe8e48fb793b8bd453db06eacf4d086e7573
Reviewed-on: http://review.couchbase.org/31619
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Trond Norbye <trond.norbye@gmail.com>