Inflate items upon retrival
If the client don't support the datatypes compressed
items should be inflated prior to returning them to the
client.

If the item is tagged as JSON we don't do any sort of
conversion of the object but just return it as it without
telling the client that we failed to transfer the information
that this is a JSON document.

Change-Id: I5c8c06c0d4890b39a832545b60bb4e8e729c656e
Reviewed-on: http://review.couchbase.org/31673
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Trond Norbye <trond.norbye@gmail.com>