MB-9693: Re-map enoent to not my vbucket for failed bg fetch
If a VBucket is deleted in the middle of a background fetch then
the result of the disk read is an enoent. When processing this
error we need to check to see if the VBucket is still valid and
also that it is in active state before returning enoent. If it's
not then we should re-map the return code to not my vbucket since
this is much more descriptive of the actual cause of the operation
failure.

Change-Id: I1131b9657a2ee7ca7990416afcd483cd9f54c4e0
Reviewed-on: http://review.couchbase.org/31832
Reviewed-by: Sundararaman Sridharan <sundar.sridharan@gmail.com>
Reviewed-by: Junyi Xie <junyi.couchbase@gmail.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>