MB-9693 Don't evict a temp item whose state is in TEMP_INIT.
If a temp item's state is TEMP_INIT, it means that we spawned a
background fetch to figure out if an item was deleted before or
never exists before. If we allow the item pager to remove this
temp item, getMeta() request can't be processed correctly and will
return an incorrect status to the client.

Change-Id: I1b025c09d22660d70dea22cb92ae728a0bb8709e
Reviewed-on: http://review.couchbase.org/32470
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Chiyoung Seo <chiyoung@couchbase.com>