Fix failing checkpoint test
The reason for this issue is that in the new checkpointing code we
don't give the checkpoint start and end messages sequence numbers.
As a result when we merge checkpoints together if a cursor is
pointing to a checkpoint start or end message we will think the
sequence number is 0 and assign the cursor to the beginning of the
checkpoint queues. This fix makes so that the checkpoint end
message has the same number as the last item +1  in the checkpoint
and also so that the checkpoint start message has the same number as
the last item in the previous checkpoint. The only time a conflict
will occur is when we merge two checkpoints together, but in this
case these items are removed so this is ok.

Change-Id: I8d011e81e1258ce26162da68a1c12626d9c1e10f
Reviewed-on: http://review.couchbase.org/32611
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>