Only create vb0 if neccesary after warmup
Since warmup will initialize a vbucket based on it's past state we
should only create the default vbucket 0 if one is not loaded
during warmup.

Note that the changes in checkpoint numbers for the tests are
needed because we are actually creating vbucket 0 properly and
that means for an active vbucket starting the open checkpoint
at 1 instead of 0.

Change-Id: I35b7d8a3b94d7292a1116496f901fb5a48634044
Reviewed-on: http://review.couchbase.org/32780
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>