Close a stream if the vbucket state changes
This change is needed for two reasons. The first is that upr
consumers should get a notification that a stream is closed if
the state of the vbucket changes. The second is that since a
vbucket might be accessed by a stream during shutdown we need
to notify the upr connection that the vbucket is no longer
valid so that the stream can be closed. Otherwise we might
access invalid objects.

Change-Id: I1ef72b7488976a058af001afb085a73cd484edc7
Reviewed-on: http://review.couchbase.org/34205
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>
Tested-by: buildbot <build@couchbase.com>