Track the number of items per vbucket with the full eviction.
As the full eviction ejects an item from cache completely, we
need a new way of tracking the number of item per vbucket when
the full eviction is chosen.

This change basically adds a new flag to the hash table item
(i.e., StoredValue) to indicate whether an item is newly added
to the cache or not, and then increments the item counter. Then,
the flusher figures out whether the counter should be decremented
because the item already exists on the DB file.

Change-Id: I492c03053238a4baa155eb76575893da323d26c4
Reviewed-on: http://review.couchbase.org/32985
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>
Tested-by: Chiyoung Seo <chiyoung@couchbase.com>