Don't use the vbucket uuid in persist seqno messages
The vbucket uuid field was added in order to make sure that things
like failover didn't happen when you issued this command. This is
unnecessary because if you want to watch seqno persistence and
there are failures then your going to have to get the latest seqno
and try again anyways.

Change-Id: If1033d03597474942216c3ca3dc4d793994d4c76
Reviewed-on: http://review.couchbase.org/33227
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>