MB-9900 Rollback implementation on a replica vbucket
This change implements the rollback support on a replica
vbucket in case where a replica vbucket is ahead of its
active vbucket.

If the number of items to be rollbacked is more than
50% of the number of items on the replica vbucket, we then
reset the replica vbucket and ask the active vbucket to send
the entire snapshot to the replica side.

Change-Id: I0fbe8d91994807073be61d65b735ca79fe6a42ba
Reviewed-on: http://review.couchbase.org/32782
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>
Tested-by: Chiyoung Seo <chiyoung@couchbase.com>