Return ENGINE_WANT_MORE when upr step should be called again
When the engine wants the memcached layer to call the step function
again then we should return ENGINE_WANT_MORE. For example, this
error code should be returned when the engine has more messages that
are ready to be written to the wire. If the engine does not have any
more messages to send then we should return ENGINE_SUCCESS. This will
cause the memcached layer to stop calling the step function. When
more messages do become ready then we need to notify the memcached
layer that more messages are ready.

There is a corresponding change in the memcached layer that allows us
to use these two error codes. This change corrects the engine code to
return the proper error codes when we merge the memcached change. A
later change will allow us to get the full behavior for pausing and
resuming the calling of the step function.

Change-Id: I5e92386327ece55bab344b8093bda4a3f1009ca4
Reviewed-on: http://review.couchbase.org/33477
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>