MB-10277 Fix to the memory corruption bug in TOUCH command
If TOUCH command is issued with the same expiration time for
a non-resident item, it causes the segmentation fault due to
the NULL pointer to an Item instance.
This change makes sure that we create an Item instance in the
above case because we still need to send the item's metadata
to the client.

Change-Id: Ic1efe24add1f0a3d2da2804c106b61c26bb275fc
Reviewed-on: http://review.couchbase.org/33848
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>
Tested-by: Chiyoung Seo <chiyoung@couchbase.com>