Add Notifiable class
We need to be able to notify the UprConsumer as well as the
UprProducer and this is different than the requirements for tap. I'm
usually not a fan of multiple inheritance, but I think in this case
it works pretty well. The idea is to have the notification state in
the Notifiable class. Then any class that wants to be notifiable in
the ConnMap just needs to extend this class. In the future when we
remove tap we can refactor this multiple inheritance out.

Change-Id: I1c8a9c19084e6d72b9710d048886f2d2243bac78
Reviewed-on: http://review.couchbase.org/34204
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>
Tested-by: buildbot <build@couchbase.com>