MB-10355: Fix issues with incorrectly setting the datatype
There were two issue here. One was that we weren't properly setting
the data type when processing an incoming tap mutation. The second
was that the buffer for meta data used when persisting items was not
zeroed out. Since the extra meta length can be zero this caused the
last byte to contain garbage. When we ended up reading this value
it caused errors with the datatype field.

Change-Id: I5162bf51c4021ac30359e3cee12184a698179201
Reviewed-on: http://review.couchbase.org/34166
Reviewed-by: abhinav dangeti <abhinav@couchbase.com>
Tested-by: abhinav dangeti <abhinav@couchbase.com>