Fix issue with removing connections in uprconnmap
There were a couple problems here. The first was that we were
removing connections incorrectly. The remove() function was causing
multiple connections with the same name to be removed. This caused
problems for any function that call findByName(). The other problem
was that since connections are freed by a background job it was
possible that the vbucket had been deleted before we freed up all
of the streams. As a result we need to close the stream before we
remove the connection from the all list.

Change-Id: I3bb62a17fc25ad27780f1c65d6d6a9a7ce4d6351
Reviewed-on: http://review.couchbase.org/34201
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>