Fix issue that caused items to be re-replicated through upr
We us the lowest sequence number in a checkpoint as the sequence
number for a checkpoint start message. During the setOpenCheckpointId
call we were resetting the sequence number to something higher and
this caused problems for sequence number replication. I also
refactored the sequence number registration so that we don't send
the item at the start sequence number.

Change-Id: Idc6555b09036c546fa58859947af09ca8100a38e
Reviewed-on: http://review.couchbase.org/34413
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: David Liao <david.liao@couchbase.com>