MB-10471: The chk low seqno should be set to the chk start item
We use the sequence number of the last item in the previous
checkpoint as the sequence number of the checkpoint start
item. As a result we should consider this number as the low
sequence number for the checkpoint we are in. This fixes errors
caused by returning -1 (the max seqno) as the sequence number
when the checkpoint didn't contain any items.

Change-Id: If33ca179e354be4d6e0a1fc43b3f8077d3b4cf00
Reviewed-on: http://review.couchbase.org/34910
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>
Tested-by: buildbot <build@couchbase.com>