MB-10644: Always try to notify an upr connection has items ready
In the current code we check to see if the ready queue is empty
before sending a notification. This can race with the actual
variable that we use to decide whether a notification needs to
be sent. Instead we should always call the notification code in
the UprConnMap because this code checks this variable and decides
whether or not to send a notification. This new check won't grab
any extra locks either.

Change-Id: Ib70e69233071562a04f1dbfc14b50bde8eb990ee
Reviewed-on: http://review.couchbase.org/34960
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>