MB-10514: The chk low seqno should be set to the chk start item
We use the sequence number of the last item in the previous
checkpoint + 1 as the sequence number of the checkpoint start
item. As a result we should consider this number as the low
sequence number for the checkpoint we are in. This fixes an error
which cause a cursor to not be registered improperly due to the
start seqno being set to -1.

Change-Id: Idd53c8a951f5844b6ac0aa1f53a78d76b9fd7e69
Reviewed-on: http://review.couchbase.org/35243
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>