MB-10772: Fix issue with vbucket getting out of sync during vb reset
During a vbucket reset which happens during a tap initial backfill
we should just delete the vbucket on disk even if there are no
items because certain data such as the vbucket high seqno and
failover table can get out of sync. Since tap will only be used for
upgrades then it should be okay to just remove the optimization of
just resetting some of the values in the vbucket without touching
disk.

Change-Id: I8b91b67bc4537c07f69b719fd464b4b20083576f
Reviewed-on: http://review.couchbase.org/35441
Tested-by: Michael Wiederhold <mike@couchbase.com>
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>